requirements.txt
uvicorn
fastapi
pydantic
sentence-transformers
openai
faiss-cpu
python-dotenv
PyPDF2
langchain